macOS version and refactor NewTaskView
