+++ Shared/TahDoodleApp.swift
@@ -12,7 +12,14 @@ struct TahDoodleApp: App {
     let taskStore = TaskStore()
     var body: some Scene {
         WindowGroup {
+            #if os(macOS)
             ContentView(taskStore: taskStore)
+                .frame(minWidth: 200,
+                       maxWidth: 300,
+                       minHeight: 200)
+            #else
+            ContentView(taskStore: taskStore)
+            #endif
         }
     }
 }
